Move ResourceAgentInfo/Type to wire-types
